gw/pyproject.toml
[project]
name = "gw"
version = "0.1.0"
description = "Gateway service for Raspberry Pi 5 using direct AWS IoT MQTT and BLE MCU communication"
readme = "README.md"
authors = [
    { name = "Maxim Parkachov", email = "mparkachov@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "bleak>=0.22,<1",
    "paho-mqtt>=2,<3",
    "watchtower>=3.3,<4",
]

[project.scripts]
gw = "gw:main"
gw-sparkplug-cmd = "gw.sparkplug_cmd:main"

[build-system]
requires = ["uv_build>=0.10.6,<0.11.0"]
build-backend = "uv_build"
